+++ nym-vpn-core/crates/nym-vpn-lib-uniffi/src/lib.rs
@@ -77,9 +77,10 @@ use sentry::ClientInitGuard;
 use tokio::{runtime::Runtime, sync::Mutex};
 
 use nym_vpn_lib_types::{
-    AccountControllerState, EntryPoint, ExitPoint, Gateway, GatewayType, Network,
-    NetworkCompatibility, ParsedAccountLinks, PrivyDerivationMessage, RegisterAccountResponse,
-    StoreAccountRequest, SystemMessage, TunnelEvent, UserAgent,
+    AccountControllerState, EntryPoint, ExitPoint, Gateway, GatewayType, GetDeeplinkParams,
+    Network, NetworkCompatibility, ParsedAccountLinks, PrivyDerivationMessage,
+    RegisterAccountResponse, StoreAccountRequest, SystemMessage, TunnelEvent, UserAgent,
+    VpnAccountSummary,
 };
 
 use account::AccountControllerHandle;
@@ -430,6 +431,20 @@ pub fn getAccountState() -> Result<AccountControllerState, VpnError> {
     RUNTIME.block_on(account::get_account_state())
 }
 
+/// Get a deeplink
+#[allow(non_snake_case)]
+#[uniffi::export]
+pub fn getDeeplink(params: GetDeeplinkParams) -> Result<String, VpnError> {
+    RUNTIME.block_on(account::get_deeplink(params))
+}
+
+/// Get the account summary
+#[allow(non_snake_case)]
+#[uniffi::export]
+pub fn getAccountSummary() -> Result<Option<VpnAccountSummary>, VpnError> {
+    RUNTIME.block_on(account::get_account_summary())
+}
+
 async fn get_account_id() -> Result<String, VpnError> {
     account::get_account_id()
         .await?
@@ -553,6 +568,7 @@ pub struct VPNConfig {
     pub exit_router: ExitPoint,
     pub enable_two_hop: bool,
     pub enable_bridges: bool,
+    pub enable_lewes_protocol: bool,
     pub residential_exit: bool,
     /// Custom DNS used when set.
     /// Leave empty to use default DNS servers.