+++ nym-vpn-core/crates/nym-vpn-proto/src/rpc_client.rs
@@ -3,8 +3,8 @@
 
 use nym_vpn_lib_types::{
     AccountBalanceResponse, AccountCommandResponse, AccountControllerState, AvailableTickets,
-    DiagnosticReport, EntryPoint, ExitPoint, FeatureFlags, Gateway, HttpRpcSettings,
-    ListGatewaysOptions, LogPath, LookupGatewayFilters, NetworkCompatibility,
+    DiagnosticReport, EntryPoint, ExitPoint, FeatureFlags, Gateway, GetDeeplinkParams,
+    HttpRpcSettings, ListGatewaysOptions, LogPath, LookupGatewayFilters, NetworkCompatibility,
     NetworkStatisticsIdentity, NymVpnDevice, NymVpnUsage, ParsedAccountLinks,
     PrivyDerivationMessage, RegistrationReport, Socks5Settings, Socks5Status, StoreAccountRequest,
     SystemMessage, TunnelEvent, TunnelState, VpnAccountSummary, VpnServiceConfig, VpnServiceInfo,
@@ -89,6 +89,15 @@ impl RpcClient {
     pub async fn set_enable_two_hop(&mut self, enable_two_hop: bool) -> Result<()> {
         self.0
             .set_enable_two_hop(enable_two_hop)
+            .await
+            .map_err(Error::Rpc)?
+            .into_inner();
+        Ok(())
+    }
+
+    pub async fn set_enable_lewes_protocol(&mut self, enable_lewes_protocol: bool) -> Result<()> {
+        self.0
+            .set_enable_lewes_protocol(enable_lewes_protocol)
             .await
             .map_err(Error::Rpc)?
             .into_inner();
@@ -525,6 +534,18 @@ impl RpcClient {
             .map_err(Error::InvalidResponse)?;
 
         Ok(account_summary)
+    }
+
+    pub async fn get_deeplink(&mut self, params: GetDeeplinkParams) -> Result<String> {
+        let request: proto::GetDeeplinkParams = params.into();
+        let url = self
+            .0
+            .get_deeplink(request)
+            .await
+            .map_err(Error::Rpc)?
+            .into_inner();
+
+        Ok(url)
     }
 
     pub async fn get_log_path(&mut self) -> Result<LogPath> {